Refactored the completion contributor:
- split up the epic methods into smaller ones.
- replaced usage of .with() by the new .withPattern() utility.
- latex and bibtex now have a separate completion contributor
- bibliographystyle completion has been moved to the Argument.Type system.
