+++ src/nl/hannahsten/texifyidea/completion/LatexListTypeProvider.kt
@@ -7,7 +7,6 @@ import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.ProcessingContext
 import com.intellij.util.containers.ContainerUtil
 import nl.hannahsten.texifyidea.TexifyIcons
-import nl.hannahsten.texifyidea.completion.handlers.MoveToEndOfCommandHandler
 
 /**
  * @author Hannah Schellekens
